Feature: use PYFILEDIR_WILDCARD to match any character except ascii
For characters in [GBK/2：OXBOA1-F7FE]
we can get frist character of pinyin easily,
for those who beyond it,we can use PYFILEDIR_WILDCARD
do wildcard matching,since I want keep the script
minimal and simple.
for a better helper would do a syntax highlight
on filename,the character in GKB/2 or not should
display with different color,
for Bash this is not possible.